requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
pandas==3.0.2
numpy==2.4.4
python-multipart==0.0.27
google-genai>=1.0.0
python-dotenv>=1.0.0
